+++ radioactive/__main__.py
@@ -102,6 +102,7 @@ def final_step(options, last_station, alias, handler):
 
     handle_listen_keypress(
         alias,
+        handler,
         player,
         target_url=options["target_url"],
         station_name=options["curr_station_name"],
@@ -110,6 +111,7 @@ def final_step(options, last_station, alias, handler):
         record_file=options["record_file"],
         record_file_format=options["record_file_format"],
         loglevel=options["loglevel"],
+        volume=options["volume"],
     )